Move escapePowerShellArg into winapp-cli-utils
Avoid a single-function file: the PowerShell-escaping helper now lives
alongside the other vscode-free CLI dispatch utility (getWinappCliPath)
in winapp-cli-utils.ts.
